+++ app/user/flask/requirements.txt
@@ -1,0 +1,3 @@
+flask
+psycopg2-binary
+Flask-SQLAlchemy
