+++ app/user/grpc_service/Dockerfile
@@ -1,0 +1,10 @@
+FROM python
+
+COPY requirements.txt .
+
+RUN pip install -r requirements.txt
+
+COPY . .
+
+EXPOSE 5001
+ENTRYPOINT ["python", "user_grpc.py"]
