+++ app/user/grpc_service/requirements.txt
@@ -1,5 +1,2 @@
-flask
-psycopg2-binary
-Flask-SQLAlchemy
 grpc-interceptor ~= 0.11.0
 grpcio-tools ~= 1.30
