+++ docker-compose.yml
@@ -6,11 +6,16 @@ services:
     build: ./app/user/database
     ports:
       - "5432:5432"
-      
+
+  user_grpc:
+    container_name: user_grpc
+    image: user_grpc
+    build: ./app/user/grpc_service
+
   user:
     container_name: user
     image: user
-    build: ./app/user/
+    build: ./app/user/flask
     ports:
       - "80:80"
     environment:
